add subscription

+++ app/api/board/route.js
@@ -25,6 +25,13 @@ export async function POST(req) {
 
     const user = await User.findById(session.user.id);
 
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first" },
+        { status: 403 },
+      );
+    }
+
     const board = await Board.create({
       userId: user._id,
       name: body.name,
@@ -58,9 +65,17 @@ export async function DELETE(req) {
       return NextResponse.json({ error: "Not authorized" }, { status: 401 });
     }
 
+    const user = await User.findById(session?.user?.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first" },
+        { status: 403 },
+      );
+    }
+
     await Board.deleteOne({ _id: boardId, userId: session?.user?.id });
 
-    const user = await User.findById(session?.user?.id);
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
     await user.save();
     return NextResponse.json({ message: "Board deleted" });
